requirements.txt
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
mysql-connector-python==8.3.0
netaddr==0.9.0
psycopg2==2.9.9
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
urllib3==2.1.0
wrapt==1.16.0
uvicorn==0.25.0
fastapi==0.108.0
pydantic==2.5.2
yookassa==3.5.0
